requirements.txt
python-dotenv
discord.py
bs4